requirements.txt.txt
asgiref                       3.8.1
channels                      4.2.0
channels_redis                4.2.1
Django                        5.1.3
django-cors-headers           4.6.0
djangorestframework           3.15.2
djangorestframework-jwt       1.11.0
djangorestframework-simplejwt 5.3.1
msgpack                       1.1.0
pillow                        11.0.0
pip                           24.2
PyJWT                         1.7.1
pyotp                         2.9.0
redis                         5.2.1
sqlparse                      0.5.1
tzdata                        2024.2